+++ HISTORY.md
@@ -4,6 +4,8 @@
 
 ### API
 
+- Add sidebar in UI (#33)
+- Show extra description of the DP (#33)
 - Make `create_figure` return a `Figure` object (#34, #35)
 - Add support for displaying multiple arrays (#23, #32)
 - Improve Front-end Callback Structure (#31)